moved domain to controller_control
