+++ app/domain/controller.py
@@ -2,7 +2,6 @@ from flask import g
 
 from app import app, db
 from app.models.controller_user import ControllerUser
-from app.models.regulation_check import UnitType, RegulationCheckType
 
 
 def create_controller_user(ac_info):
@@ -38,17 +37,3 @@ def get_controller_from_ac_info(ac_info):
     return ControllerUser.query.filter(
         ControllerUser.agent_connect_id == agent_connect_id
     ).one_or_none()
-
-
-def get_no_lic_observed_infractions(control_date):
-    return [
-        {
-            "sanction": "NATINF 23103",
-            "date": control_date.isoformat(),
-            "is_reportable": True,
-            "is_reported": True,
-            "extra": None,
-            "check_unit": UnitType.DAY,
-            "check_type": RegulationCheckType.NO_LIC,
-        }
-    ]
